+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 # phpstorm-xdebug-skip Changelog
 
+## 0.1.0 - 2021-10-17
+### Added
+- Initial release
+- Use stepinto to exit whereever the breakpoint resolves to a non match.
+- Added settings panel to configure the plugin
+- Namespace and filepath selector for paths to skip when debugging
+- Added extra text as the namespace selector doesn't appear on the first click on +
 ## [Unreleased]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)